feat(turn): pool envelope carries the hosted workspace/user turn context
The gateway's mission preparer injects the org's shared note and the caller's
context into every message-send; the pool envelope now carries the same two
strings and the per-turn session overlays them exactly like the long-lived
server path (either present = use these). Also splits parseTurnRequest and the
pi-turn projection into their own modules to keep the turn files under the
size limit.
